refactor main editor
