+++ client/src/app/notion-like/page.tsx
@@ -1,10 +1,10 @@
-import { NotionEditor } from "@/components/tiptap-templates/notion-like/notion-like-editor"
+import { NotionEditor } from "@/components/tiptap-templates/notion-like/notion-like-editor";
 
 export default async function Page({
   params,
 }: {
-  params: Promise<{ slug: string }>
+  params: Promise<{ slug: string }>;
 }) {
-  const slug = (await params).slug || "notion"
-  return <NotionEditor room={slug} />
+  const slug = (await params).slug || "notion";
+  return <NotionEditor room={slug} />;
 }
