+++ client/src/hooks/useAutoResize.ts
@@ -1,0 +1,14 @@
+import { useEffect, useRef } from "react";
+
+export const useAutoResize = (value: string) => {
+    const textareaRef = useRef<HTMLTextAreaElement>(null);
+  
+    useEffect(() => {
+      if (textareaRef.current) {
+        textareaRef.current.style.height = "auto";
+        textareaRef.current.style.height = `${textareaRef.current.scrollHeight}px`;
+      }
+    }, [value]);
+  
+    return textareaRef;
+  };
