+++ hasura-metadata/metadata/databases/default/tables/public_blocks.yaml
@@ -114,3 +114,8 @@ update_permissions:
       filter: {}
       check: null
     comment: ""
+delete_permissions:
+  - role: user
+    permission:
+      filter: {}
+    comment: ""
